requirements.txt
# Civil Engineering + AI News Collection Pipeline
# Requirements

# HTTP requests
requests>=2.31.0

# RSS/Atom parsing
feedparser>=6.0.10

# HTML parsing
beautifulsoup4>=4.12.0
lxml>=5.0.0

# YAML config
PyYAML>=6.0.1

# Date parsing
python-dateutil>=2.8.2
